Organized the files and begin to add tests for the gem
