+++ .gitignore
@@ -1,5 +1,3 @@
-*.gemspec
-
-pkg/**/*
+pkg
 nbproject/**/*
-
+.DS_Store
