+++ Gemfile
@@ -1,0 +1,3 @@
+#!/usr/bin/env ruby
+
+gem 'infinity_test'
