--- lib/html_and_aml/html_and_aml.rb
@@ -1,10 +1,0 @@
-
-require File.join(File.dirname(__FILE__), "helpers/entities")
-require File.join(File.dirname(__FILE__), "helpers/tags")
-
-module HtmlAndAML # == HTML And Another Markup Language
-
-  include Entities
-  include Tags
-
-end
