+++ lib/string_utility_belt/general.rb
@@ -13,7 +13,7 @@ module General
     self.strip.gsub!(/\s+/, " ")
   end
 
-  def have_this_words? words_to_match, exact_word=false
+  def have_this_words?(words_to_match, exact_word=false)
     helper_have_this_words? words_to_match, exact_word do |string, word, exact_world|
       return false if (string !~ word.regex_me_to_search_ruby(:exact_word => exact_word, :case_insensitive => true))
     end
