+++ lib/string_utility_belt/html_and_aml.rb
@@ -1,0 +1,7 @@
+
+module HtmlAndAML # == HTML And Another Markup Language
+
+  include Entities
+  include Tags
+
+end
