+++ lib/string_utility_belt/match_rank.rb
