+++ lib/string_utility_belt/regex_me_helper.rb
@@ -1,3 +1,4 @@
+# coding: utf-8
 
 class String
