+++ test/string_utility_belt/entities_test.rb
@@ -1,0 +1,1 @@
+require 'test_helper'
