+++ test/string_utility_belt/html_and_aml_test.rb
@@ -1,0 +1,2 @@
+require 'test_helper'
+
