+++ test/test_helper.rb
@@ -1,0 +1,14 @@
+
+# To run tests: 
+#
+# 1) Install infinity_test
+#
+# gem install infinity_test
+#
+# 2) Run the tests:
+#
+# infinity_test --test-unit
+#
+
+require 'test/unit'
+require 'string_utility_belt'
